requirements.txt
fastapi
uvicorn[standard]
anthropic
openai
python-dotenv
pydantic
